requirements.txt
cloudinary==1.36.0
Django==4.2.1
djangorestframework==3.14.0
langchain==0.0.312
python-dotenv==1.0.0
pypdf==3.16.4
openai==0.28.1
tiktoken==0.5.1
faiss-cpu==1.7.4